// utils/enums.ts
export enum Technology {
  CPlusPlus = "C++",
  Python = "Python",
  JavaScript = "JavaScript",
  TypeScript = "TypeScript",
  GoLang = "GoLang",
  React = "React",
  Angular = "Angular",
  FastAPI = "FastAPI",
  MongoDB = "MongoDB",
  Docker = "Docker",
  Kubernetes = "Kubernetes",
  Bash = "Bash",
  Kotlin = "Kotlin",
  JetpackCompose = "Jetpack Compose",
  Flask = "Flask",
  PostgreSQL = "PostgreSQL",
  Firebase = "Firebase",
  NodeJS = "NodeJS",
  ExpressJS = "ExpressJS",
  Redis = "Redis",
  Flutter = "Flutter",
  Go = "Go",
  GinGonic = "GinGonic",
}